+++ src/code.hpp
@@ -78,7 +78,6 @@ class Code
     }
 
     std::size_t size();
-    std::size_t &nested_scopes();
     void push_break(std::size_t ind);
     void set_break_to(std::size_t ind, std::size_t base);
     void push_continue(std::size_t ind);